move retry func to public
