+++ pkg/kube/structured/structured_helper.go
@@ -19,7 +19,7 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/keikoproj/kubedog/internal/util"
+	"github.com/keikoproj/kubedog/pkg/util"
 	"github.com/keikoproj/kubedog/pkg/kube/common"
 	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
