+++ pkg/kube/unstructured/unstructured.go
@@ -24,8 +24,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/keikoproj/kubedog/internal/util"
 	"github.com/keikoproj/kubedog/pkg/kube/common"
+	"github.com/keikoproj/kubedog/pkg/util"
 	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
 	"golang.org/x/text/cases"
