+++ pkg/kube/unstructured/unstructured_helper.go
@@ -21,7 +21,7 @@ import (
 	"html/template"
 	"os"
 
-	"github.com/keikoproj/kubedog/internal/util"
+	"github.com/keikoproj/kubedog/pkg/util"
 	"github.com/pkg/errors"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
